feat: rename core.ts to internal.ts
