+++ src/browser.ts
@@ -1,3 +1,3 @@
-import { Eta as EtaCore } from "./core.ts";
+import { Eta as EtaCore } from "./internal.ts";
 
 export class Eta extends EtaCore {}
