+++ src/file-handling.ts
@@ -2,7 +2,7 @@ import * as fs from "node:fs";
 import * as path from "node:path";
 
 import type { Options } from "./config.ts";
-import type { Eta as EtaCore } from "./core.ts";
+import type { Eta as EtaCore } from "./internal.ts";
 import { EtaFileResolutionError } from "./err.ts";
 
 export function readFile(this: EtaCore, path: string): string {
